pyproject.toml
[tool.poetry]
name = "handyman"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9"
Django = "^3.1.6"
django-stubs-ext = "^0.1.0"
psycopg2-binary = "^2.8.6"
python-dotenv = "^0.15.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.0.0"}
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
mypy = "^0.800"
black = "^20.8b1"
flake8 = "^3.8.4"
pytest = "^6.2.2"
django-stubs = "^1.7.0"
isort = "^5.7.0"
django-coverage-plugin = "^1.8.0"
livereload = "^2.6.3"
whitenoise = "^5.2.0"
django-split-settings = "^1.0.1"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
pytest-deadfixtures = "^2.2.1"
pytest-timeout = "^1.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
